The issue with connector to polyline is now solved.
